ManufactureAPI_v01/requirements.txt
pymongo
fastapi[all] 
python-dotenv
pydantic-settings
PyJWT
python-decouple